[refactor][step3] UI로직 분리 및 클래스 명칭 _ 제거
