+++ src/test/java/step3RacingCarCompetition/RacingCarCompetitionTest.java
@@ -1,12 +1,14 @@
-package step3_RacingCar_Competition;
+package step3RacingCarCompetition;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -20,7 +22,7 @@ class RacingCarCompetitionTest {
     @BeforeAll
     static void init() {
         racingCarCompetition = new RacingCarCompetition();
-        racingCarInfo = new RacingCarInfo(totalCarNumber, totalMovingCount);
+        racingCarInfo = new RacingCarInfo(List.of(totalCarNumber, totalMovingCount));
         customMovingData = new ArrayList<>();
         ArrayList<Integer> innerData = new ArrayList<>();
         innerData.add(1);
@@ -32,7 +34,7 @@ class RacingCarCompetitionTest {
         innerData.add(4);
         customMovingData.add(innerData);
 
-        racingCarCompetition.recordRacingCarData(customMovingData, racingCarInfo);
+
     }
 
     @DisplayName(value = "랜덤값 범위 확인 테스트")
@@ -43,17 +45,16 @@ class RacingCarCompetitionTest {
 
     @DisplayName(value = "이동, 정지 테스트")
     @ParameterizedTest()
-    @CsvSource({"0,false", "1,false", "2,false", "3,false", "4,true", "5,true",
-            "6,true", "7,true", "8,true", "9,true"})
+    @CsvSource({"0,false", "5,true", "9,true"})
     void movingStopDecisionTest(int randomNumber, boolean isGoStop) {
         assertThat(racingCarCompetition.moveStopDecision(randomNumber)).isEqualTo(isGoStop);
     }
 
+    @Test
     @DisplayName(value = "차 이동상태 저장 여부 테스트, 차2대,이동 2번")
-    @ParameterizedTest()
-    @CsvSource({"0,-", "1,--"})
-    void recordRacingCarDataTest(int carNumber, String carMovingString) {
-        assertThat(racingCarInfo.findRacingCarData(carNumber)).isEqualTo(carMovingString);
+    void recordRacingCarDataTest() {
+        racingCarCompetition.recordRacingCarData(customMovingData, racingCarInfo);
+        assertThat(racingCarInfo.findRacingCarData(1)).isEqualTo("--");
     }
